% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_example_documentation.R
\docType{data}
\name{cov_example}
\alias{cov_example}
\title{Covariate Information}
\format{
A dataframe with 500 rows and 4 variables:
\describe{
 \item{y}{phenotypic factor of interest}
 \item{BMI}{BMI in kg/m^2}
 \item{age}{age in years}
 \item{sex}{sex coded as 1 & 2}
 }
}
\usage{
cov_example
}
\description{
Dataset containing example covariate information for analysis
}
\keyword{datasets}
