% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_example_documentation.R
\docType{data}
\name{geno_example}
\alias{geno_example}
\title{Genotypic Information}
\format{
An object of class \code{matrix} with 500 rows and 20 columns.
}
\usage{
geno_example
}
\description{
Dataset containing example genotypic information for analysis, coded as 0 1 2. Has 500 observations
and assumes a minor allele frequency threshold of 0.05. We included the first 20 SNPs to meet this
threshold in this example. Each column holds the information for an individual SNP, and each row
contains the genotypic information for that observation.
}
\keyword{datasets}
